requirements.txt
pyautogen
python-dotenv==1.0.1
pytest==8.2.2
gitpython==3.1.43
ordered-set==4.1.0
PyGithub==2.3.0
chardet==5.2.0
markdown==3.6
langchain
langchain-google-genai
langchain-community
langchain-core
langchain-text-splitters
langchain-ollama
faiss-cpu
GitPython
